information_system_gkh/requirements.txt
celery[redis]==5.3.6
flower==1.2.0
Django==4.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
